requirements.txt
aiohttp==3.11.7
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
async-fastapi-jwt-auth==0.6.6
asyncpg==0.30.0
backoff==2.2.1
certifi==2024.12.14
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isort==5.13.2
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
opentelemetry-api==1.29.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-sdk==1.29.0
orjson==3.10.14
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.19
redis==5.2.1
ruff==0.8.2
setuptools==75.6.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
